add pollFragment to TabView
