+++ app/src/main/java/com/example/unserhoersaal/views/PollFragment.java
@@ -1,38 +1,59 @@
 package com.example.unserhoersaal.views;
 
 import android.os.Bundle;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.databinding.DataBindingUtil;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
 import com.example.unserhoersaal.R;
+import com.example.unserhoersaal.databinding.FragmentPollBinding;
+import com.example.unserhoersaal.viewmodel.PollViewModel;
 
-/**Polls.*/
+/** View, which displays the voting during a meeting. */
 public class PollFragment extends Fragment {
+
+  private static final String TAG = "PollFragment";
+
+  private FragmentPollBinding binding;
+  private PollViewModel pollViewModel;
 
   public PollFragment() {
     // Required empty public constructor
   }
 
   @Override
-  public void onCreate(Bundle savedInstanceState) {
+  public void onCreate(@Nullable Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
   }
 
+  @Nullable
   @Override
-  public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                           Bundle savedInstanceState) {
-    // Inflate the layout for this fragment
-    return inflater.inflate(R.layout.fragment_poll, container, false);
+  public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
+                           @Nullable Bundle savedInstanceState) {
+    this.binding = DataBindingUtil.inflate(inflater,
+            R.layout.fragment_poll, container, false);
+    return this.binding.getRoot();
   }
 
   @Override
   public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
     super.onViewCreated(view, savedInstanceState);
+
+    this.initViewModel();
+    this.connectBinding();
+  }
+
+  private void initViewModel() {
+    this.pollViewModel = new ViewModelProvider(requireActivity())
+            .get(PollViewModel.class);
+    this.pollViewModel.init();
+  }
+
+  private void connectBinding() {
+    this.binding.setLifecycleOwner(getViewLifecycleOwner());
   }
 }
