--- app/src/main/java/com/example/unserhoersaal/views/VotingFragment.java
@@ -1,59 +1,0 @@
-package com.example.unserhoersaal.views;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.databinding.DataBindingUtil;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.ViewModelProvider;
-import com.example.unserhoersaal.R;
-import com.example.unserhoersaal.databinding.FragmentVotingBinding;
-import com.example.unserhoersaal.viewmodel.VotingViewModel;
-
-/** View, which displays the voting during a meeting. */
-public class VotingFragment extends Fragment {
-
-  private static final String TAG = "VotingFragment";
-
-  private FragmentVotingBinding binding;
-  private VotingViewModel votingViewModel;
-
-  public VotingFragment() {
-    // Required empty public constructor
-  }
-
-  @Override
-  public void onCreate(@Nullable Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-  }
-
-  @Nullable
-  @Override
-  public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container,
-                           @Nullable Bundle savedInstanceState) {
-    this.binding = DataBindingUtil.inflate(inflater,
-            R.layout.fragment_voting, container, false);
-    return this.binding.getRoot();
-  }
-
-  @Override
-  public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-    super.onViewCreated(view, savedInstanceState);
-
-    this.initViewModel();
-    this.connectBinding();
-  }
-
-  private void initViewModel() {
-    this.votingViewModel = new ViewModelProvider(requireActivity())
-            .get(VotingViewModel.class);
-    this.votingViewModel.init();
-  }
-
-  private void connectBinding() {
-    this.binding.setLifecycleOwner(getViewLifecycleOwner());
-  }
-}
