+++ app/src/main/res/layout/fragment_poll.xml
@@ -1,10 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <data>
-
+        <import type="com.example.unserhoersaal.utils.NavUtil"/>
+        <import type="com.example.unserhoersaal.R"/>
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -86,7 +87,7 @@
                         android:layout_alignParentTop="true"
                         android:layout_marginBottom="8dp"
                         android:nestedScrollingEnabled="false"
-                        tools:listitem="@layout/poll_item"/>
+                        tools:listitem="@layout/poll_item" />
                 </RelativeLayout>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
@@ -103,7 +104,8 @@
             app:tint="@color/white"
             style="@style/fabStyle"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent" />
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:onClick="@{(view) -> NavUtil.navigate(view, R.id.action_courseMeetingFragment_to_createPollFragment)}"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
