+++ app/src/main/res/navigation/nav_graph.xml
@@ -296,6 +296,9 @@
         <action
             android:id="@+id/action_courseMeetingFragment_to_createThreadFragment"
             app:destination="@id/createThreadFragment" />
+        <action
+            android:id="@+id/action_courseMeetingFragment_to_createPollFragment"
+            app:destination="@id/createPollFragment" />
     </fragment>
     <fragment
         android:id="@+id/editProfileNameFragment"
@@ -475,5 +478,19 @@
         android:id="@+id/createPollFragment"
         android:name="com.example.unserhoersaal.views.CreatePollFragment"
         android:label="CreatePollFragment"
-        tools:layout="@layout/fragment_create_poll"/>
+        tools:layout="@layout/fragment_create_poll">
+        <action
+            android:id="@+id/action_createPollFragment_to_courseMeetingFragment"
+            app:destination="@id/courseMeetingFragment"
+            app:popUpTo="@id/courseMeetingFragment"
+            app:popUpToInclusive="true"/>
+    </fragment>
+    <fragment
+        android:id="@+id/pollFragment"
+        android:name="com.example.unserhoersaal.views.PollFragment"
+        android:label="PollFragment" >
+        <action
+            android:id="@+id/createPollFragment"
+            app:destination="@id/createPollFragment" />
+    </fragment>
 </navigation>
